Rename agent.py to base_agent.py and update imports
