+++ sgr_deep_research/core/agents/__init__.py
@@ -1,6 +1,6 @@
 """Agents module for SGR Deep Research."""
 
-from sgr_deep_research.core.agents.agent import BaseAgent
+from sgr_deep_research.core.agents.base_agent import BaseAgent
 from sgr_deep_research.core.agents.sgr_agent import SGRResearchAgent
 from sgr_deep_research.core.agents.sgr_auto_tools_agent import SGRAutoToolCallingResearchAgent
 from sgr_deep_research.core.agents.sgr_so_tools_agent import SGRSOToolCallingResearchAgent
