+++ sgr_deep_research/core/agents/base_agent.py
