+++ sgr_deep_research/core/agents/sgr_agent.py
@@ -2,7 +2,7 @@ import logging
 import uuid
 from typing import Type
 
-from sgr_deep_research.core.agents.agent import BaseAgent
+from sgr_deep_research.core.agents.base_agent import BaseAgent
 from sgr_deep_research.core.tools import (
     AgentCompletionTool,
     ClarificationTool,
