+++ sgr_deep_research/core/agents/tools_agent.py
@@ -5,7 +5,7 @@ from typing import Literal, Type
 from openai import pydantic_function_tool
 from openai.types.chat import ChatCompletionFunctionToolParam
 
-from sgr_deep_research.core.agents.agent import BaseAgent
+from sgr_deep_research.core.agents.base_agent import BaseAgent
 from sgr_deep_research.core.tools import (
     AgentCompletionTool,
     ClarificationTool,
